+++ ms-compras/src/main/java/com/frankester/mscompras/models/dto/CarritoResponse.java
@@ -2,11 +2,12 @@ package com.frankester.mscompras.models.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class CarritoResponse {
     private Long compradorId;
     private List<ItemDTO> items;
-    private Integer precioCarrito;
+    private BigDecimal precioCarrito;
 }
